streamlit/requirements.txt
accelerate
bitsandbytes
trafilatura
streamlit
pymupdf
pymupdf4llm
datasets
haystack-ai
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
torchaudio --index-url https://download.pytorch.org/whl/cpu
tensorflow
tokenizers
transformers
huggingface-hub
sentence-transformers
openai


















